app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
google-cloud-secret-manager>=2.18.0
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-instrumentation-fastapi>=0.44b0
python-json-logger>=2.0.7
